+++ src/main.js
@@ -1,5 +1,8 @@
 import { CSS_HIDE_OVERFLOW_CLASS, CSS_FILMS_CONTAINER_CLASS } from './constants.js';
-import { RenderPosition, getNode, renderElement, renderElements, sortArrayOfObjects } from './util.js';
+import { RenderPosition, renderElement, renderElements } from './utils/render.js';
+import { getNode } from './utils/nodes.js';
+import { sortArrayOfObjects } from './utils/common.js';
+
 import RatingView from './view/rating.js';
 import MenuView from './view/menu.js';
 import FilterView from './view/filter.js';
@@ -10,6 +13,7 @@ import EmptyFilmListView from './view/emptyFilmList.js';
 import FilmCardView from './view/filmCard.js';
 import ShowMoreView from './view/showMoreButton';
 import FilmDetailsView from './view/filmDetails.js';
+import CommentView from './view/comment.js';
 import TotalFilmsNumView from './view/totalFilmsNum.js';
 
 import { generateFilm } from './mock/film.js';
@@ -47,25 +51,25 @@ const bodyNode = getNode('body');
 // ---------------------------------------------------------
 
 // рейтинг пользователя
-renderElement(headerNode, new RatingView(userStats.rank).getElement());
+renderElement(headerNode, new RatingView(userStats.rank));
 
 // сортировка
-renderElement(mainNode, new SortsView().getElement(), RenderPosition.AFTER_BEGIN);
+renderElement(mainNode, new SortsView(), RenderPosition.AFTER_BEGIN);
 
 // меню с фильтрами
 const menuComponent = new MenuView();
-renderElement(mainNode, menuComponent.getElement(), RenderPosition.AFTER_BEGIN);
+renderElement(mainNode, menuComponent, RenderPosition.AFTER_BEGIN);
 
 // фильтры в меню
 const filtersContainer = getNode('.main-navigation__items');
-const filters = filtersData.map((filterData) => new FilterView(filterData).getElement());
+const filters = filtersData.map((filterData) => new FilterView(filterData));
 renderElements(filtersContainer, filters);
 
 // общее количество фильмов
-renderElement(footerStatsNode, new TotalFilmsNumView().getElement());
+renderElement(footerStatsNode, new TotalFilmsNumView());
 
 // статистика пользователя
-// renderElement(mainNode, new StatsView(userStats).getElement(), RenderPosition.AFTER_BEGIN);
+// renderElement(mainNode, new StatsView(userStats), RenderPosition.AFTER_BEGIN);
 
 // ---------------------------------------------------------
 // список фильмов
@@ -74,15 +78,15 @@ let allFilmsContainer;
 
 if (filmsData.length > 0) {
   allFilmListComponent = new FilmListView();
-  renderElement(filmsNode, allFilmListComponent.getElement());
+  renderElement(filmsNode, allFilmListComponent);
 
   // рендер фильмов в списке
-  allFilmsContainer = getNode(`.${CSS_FILMS_CONTAINER_CLASS}`, allFilmListComponent.getElement());
-  const films = filmsData.slice(0, FILM_COUNT_PER_STEP).map((filmData) => new FilmCardView(filmData).getElement());
+  allFilmsContainer = getNode(`.${CSS_FILMS_CONTAINER_CLASS}`, allFilmListComponent);
+  const films = filmsData.slice(0, FILM_COUNT_PER_STEP).map((filmData) => new FilmCardView(filmData));
 
   renderElements(allFilmsContainer, films);
 } else {
-  renderElement(filmsNode, new EmptyFilmListView().getElement());
+  renderElement(filmsNode, new EmptyFilmListView());
 }
 
 // список фильмов top rated
@@ -90,13 +94,13 @@ let topRatedListComponent;
 
 if (filmsData.length > 0) {
   topRatedListComponent = new FilmListView(true, 'TOP_RATED');
-  renderElement(filmsNode, topRatedListComponent.getElement());
+  renderElement(filmsNode, topRatedListComponent);
 
   // рендер фильмов в списке
-  const topFilmsContainer = getNode(`.${CSS_FILMS_CONTAINER_CLASS}`, topRatedListComponent.getElement());
+  const topFilmsContainer = getNode(`.${CSS_FILMS_CONTAINER_CLASS}`, topRatedListComponent);
   const topRatedFilms = sortArrayOfObjects(filmsData, 'rating')
     .slice(0, TOP_RATED_MAX)
-    .map((filmData) => new FilmCardView(filmData).getElement());
+    .map((filmData) => new FilmCardView(filmData));
 
   renderElements(topFilmsContainer, topRatedFilms);
 }
@@ -106,13 +110,13 @@ let mostCommentedListComponent;
 
 if (filmsData.length > 0) {
   mostCommentedListComponent = new FilmListView(true, 'MOST_COMMENTED');
-  renderElement(filmsNode, mostCommentedListComponent.getElement());
+  renderElement(filmsNode, mostCommentedListComponent);
 
   // рендер фильмов в списке
-  const mostCommentedFilmsContainer = getNode(`.${CSS_FILMS_CONTAINER_CLASS}`, mostCommentedListComponent.getElement());
+  const mostCommentedFilmsContainer = getNode(`.${CSS_FILMS_CONTAINER_CLASS}`, mostCommentedListComponent);
   const mostCommentedFilms = sortArrayOfObjects(filmsData, 'commentNumber')
     .slice(0, MOST_COMMENTED_MAX)
-    .map((filmData) => new FilmCardView(filmData).getElement());
+    .map((filmData) => new FilmCardView(filmData));
 
   renderElements(mostCommentedFilmsContainer, mostCommentedFilms);
 }
@@ -120,8 +124,8 @@ if (filmsData.length > 0) {
 // ---------------------------------------------------------
 // обработчик кликов на карточки фильмов
 const onFilmListClick = (evt) => {
-  const target = evt.target;
   evt.preventDefault();
+  const target = evt.target;
 
   if (!target.matches('.film-card__title') && !target.matches('.film-card__poster') && !target.matches('.film-card__comments')) {
     return;
@@ -130,17 +134,37 @@ const onFilmListClick = (evt) => {
   const filmCard = target.closest('.film-card');
 
   let filmId = null;
-  if (target) {
+  if (filmCard) {
     filmId = Number(filmCard.getAttribute('data-film-id'));
   }
 
+  // html элемент не содержит аттрибут с id фильма
+  if (filmId === null) {
+    return;
+  }
+
+  // найдём фильм по id в наших данных фильмов
+  const filmData = filmsData.find((item) => item.filmId === filmId);
+
+  // фильм не найден в данных, что-то не так
+  if (filmData === undefined) {
+    return;
+  }
+
   // детальная информация о фильме
-  let popupComponent = new FilmDetailsView(filmsData.find((item) => item.filmId === filmId));
-  renderElement(bodyNode, popupComponent.getElement());
+  let popupComponent = new FilmDetailsView(filmData);
+  renderElement(bodyNode, popupComponent);
+
+  // комментарии к фильму
+  const commentsContainer = getNode('.film-details__comments-list', popupComponent);
+  const commentComponents = filmData.comments.map((commentData) => new CommentView(commentData));
+
+  renderElements(commentsContainer, commentComponents);
 
   // дет. информация отрендерена, надо убрать полосу прокрутки у body
   bodyNode.classList.add(CSS_HIDE_OVERFLOW_CLASS);
 
+  // повесим обработчик закрытия
   const popupFilmCloseButton = getNode(`.film-details[data-film-id="${filmId}"] .film-details__close-btn`);
 
   const onCloseButtonClick = () => {
@@ -174,7 +198,7 @@ const onFilmListClick = (evt) => {
 // навешиваем обработки на списки фильмов
 [allFilmListComponent, topRatedListComponent, mostCommentedListComponent].forEach((fimlListComponent) => {
   if (fimlListComponent) {
-    fimlListComponent.getElement().addEventListener('click', onFilmListClick);
+    fimlListComponent.setClickHandler(onFilmListClick);
   }
 });
 
@@ -182,8 +206,7 @@ const onFilmListClick = (evt) => {
 // кнопка показать больше
 if (filmsData.length > FILM_COUNT_PER_STEP) {
   let showMoreComponent = new ShowMoreView();
-  const showMoreButton = showMoreComponent.getElement();
-  renderElement(allFilmListComponent.getElement(), showMoreComponent.getElement());
+  renderElement(allFilmListComponent, showMoreComponent);
 
   let renderElementedFilmsCount = FILM_COUNT_PER_STEP;
 
@@ -193,7 +216,7 @@ if (filmsData.length > FILM_COUNT_PER_STEP) {
     // рендер ешё группы фильмов в списке
     const films = filmsData
       .slice(renderElementedFilmsCount, renderElementedFilmsCount + FILM_COUNT_PER_STEP)
-      .map((filmData) => new FilmCardView(filmData).getElement());
+      .map((filmData) => new FilmCardView(filmData));
 
     renderElements(allFilmsContainer, films);
 
@@ -202,12 +225,12 @@ if (filmsData.length > FILM_COUNT_PER_STEP) {
 
     // если больше нечего рендерить, то убираем хендлеры и удаляем кнопку 'Show-more'
     if (renderElementedFilmsCount >= filmsData.length) {
-      showMoreButton.removeEventListener('click', onShowMoreClick);
+      showMoreComponent.removeClickHandler();
 
       showMoreComponent.removeElement();
       showMoreComponent = null;
     }
   };
 
-  showMoreButton.addEventListener('click', onShowMoreClick);
+  showMoreComponent.setClickHandler(onShowMoreClick);
 }
