+++ src/view/filter.js
@@ -1,4 +1,5 @@
-import { createElementFromTemplate, capitalizeFirstLetter } from '../util.js';
+import ComponentView from './abstract-component.js';
+import { capitalizeFirstLetter } from '../utils/common.js';
 
 // ---------------------------------------------------------
 const getFilterTemplate = ({ filterName, count }) => {
@@ -7,24 +8,13 @@ const getFilterTemplate = ({ filterName, count }) => {
 };
 
 // ---------------------------------------------------------
-export default class Filter {
+export default class Filter extends ComponentView {
   constructor(data) {
+    super();
     this._data = data;
   }
 
   getTemplate() {
     return getFilterTemplate(this._data);
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElementFromTemplate(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
   }
 }
